+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## v0.5.0
+
+p5.svg@0.5.0 (requires p5.js@0.4.8)
+
+- Fix loadSVG in p5.js@0.4.8
+
+- RendererSVG.prototype.appendChild(SVGElement)
+
+- Configure ESLint and make it happy
+
+- Allow save SVGElement in save() and saveSVG()
+
+- add p5.prototype.registerSVGFilter
+
+- add p5.prototype.getDataURL
+
 ## v0.4.3
 
 p5.svg@0.4.3 (requires p5.js@0.4.7 with [patch#850](https://github.com/processing/p5.js/pull/850))